+++ QuickStart.WepApi/Controllers/NotificationTypeController.cs
@@ -1,6 +1,6 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using QuickStart.WepApi.Context;
+using QuickStart.WepApi.DTOs.NotificationTypeDTOs;
 using QuickStart.WepApi.Entities;
 using QuickStart.WepApi.Entity;
 
@@ -20,44 +20,65 @@ namespace QuickStart.WepApi.Controllers
         [HttpGet]
         public IActionResult NotificationTypeList()
         {
-            var values = _context.NotificationTypes.ToList();
+            var values = _context.NotificationTypes
+                .Select(x => new ResultNotificationTypeDto
+                {
+                    NotificationTypeId = x.NotificationTypeId,
+                    Name = x.Name
+                })
+                .ToList();
             return Ok(values);
-        }
-
-        [HttpGet("NotificationTypeCount")]
-        public IActionResult NotificationTypeCount()
-        {
-            var value = _context.NotificationTypes.Count();
-            return Ok(value);
         }
 
         [HttpGet("{id}")]
         public IActionResult GetById(int id)
         {
-            var notificationType = _context.NotificationTypes.Find(id);
-            return Ok(notificationType);
+            var value = _context.NotificationTypes.Find(id);
+            if (value == null)
+                return NotFound();
+
+            var result = new ResultNotificationTypeDto
+            {
+                NotificationTypeId = value.NotificationTypeId,
+                Name = value.Name
+            };
+            return Ok(result);
         }
 
         [HttpPost]
-        public IActionResult CreateNotificationType(NotificationType notificationType)
+        public IActionResult CreateNotificationType(CreateNotificationTypeDto createDto)
         {
+            var notificationType = new NotificationType
+            {
+                Name = createDto.Name
+            };
+
             _context.NotificationTypes.Add(notificationType);
             _context.SaveChanges();
             return Ok("Ekleme işlemi başarı ile gerçekleşti");
         }
 
         [HttpPut]
-        public IActionResult UpdateNotificationType(NotificationType notificationType)
+        public IActionResult UpdateNotificationType(UpdateNotificationTypeDto updateDto)
         {
+            var notificationType = new NotificationType
+            {
+                NotificationTypeId = updateDto.NotificationTypeId,
+                Name = updateDto.Name
+            };
+
             _context.NotificationTypes.Update(notificationType);
             _context.SaveChanges();
             return Ok("Güncelleme işlemi başarı ile gerçekleşti");
         }
 
-        [HttpDelete]
+        [HttpDelete("{id}")]
         public IActionResult DeleteNotificationType(int id)
         {
             var value = _context.NotificationTypes.Find(id);
+            if (value == null)
+                return NotFound();
+
             _context.NotificationTypes.Remove(value);
             _context.SaveChanges();
             return Ok("Silme işlemi başarı ile gerçekleşti");
